backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.21
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
httpx>=0.25.0
pytest>=7.4.2
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0 